zen_stage1_fix/requirements.txt
pandas>=2.2.2
yfinance>=0.2.52
snowflake-connector-python>=3.11.0
python-dateutil>=2.9.0.post0
pytz>=2024.1
requests>=2.32.3